+++ abstractions/ruby/microsoft_kiota_abstractions/lib/microsoft_kiota_abstractions/serialization/parse_node.rb
@@ -52,6 +52,10 @@ module MicrosoftKiotaAbstractions
     def get_enum_value(type)
       raise NotImplementedError.new
     end
+
+    def get_child_node(name)
+      raise NotImplementedError.new
+    end
     
   end
 end