+++ http/ruby/faraday/microsoft_kiota_faraday/spec/parameters_name_decoding_handler_spec.rb
@@ -20,29 +20,6 @@ RSpec.describe MicrosoftKiotaFaraday do
         end
     end
 
-    it "doesn't decode when disabled on handler" do
-        values = Hash.new
-        values["?%24select=diplayName&api%2Dversion=2"] = "?%24select=diplayName&api%2Dversion=2"
-        values["?%24select=diplayName&api%7Eversion=2"] = "?%24select=diplayName&api%7Eversion=2"
-        values["?%24select=diplayName&api%2Eversion=2"] = "?%24select=diplayName&api%2Eversion=2"
-        values["/api-version/?%24select=diplayName&api%2Eversion=2"] = "/api-version/?%24select=diplayName&api%2Eversion=2"
-        values[""] = ""
-
-        option = MicrosoftKiotaFaraday::Middleware::ParametersNameDecodingOption.new(false)
-        handler = MicrosoftKiotaFaraday::Middleware::ParametersNameDecodingHandler.new(nil, option)
-        values.each do |key, value|
-            env = {
-                url: URI.parse("https://graph.microsoft.com/v1.0/users#{key}"),
-                request: {
-                    context: {
-                    }
-                }
-            }
-            handler.call(env)
-            expect(env[:url].to_s).to eq("https://graph.microsoft.com/v1.0/users#{value}")
-        end
-    end
-
     it "doesn't decode when disabled on request" do
         values = Hash.new
         values["?%24select=diplayName&api%2Dversion=2"] = "?%24select=diplayName&api%2Dversion=2"