+++ src/Kiota.Builder/PathSegmenters/RubyPathSegmenter.cs
@@ -4,20 +4,19 @@ using System.Linq;
 using Kiota.Builder.CodeDOM;
 using Kiota.Builder.Extensions;
 
-namespace Kiota.Builder.PathSegmenters {
-    public class RubyPathSegmenter : CommonPathSegmenter
+namespace Kiota.Builder.PathSegmenters;
+public class RubyPathSegmenter : CommonPathSegmenter
+{
+    public RubyPathSegmenter(string rootPath, string clientNamespaceName) : base(rootPath, clientNamespaceName) { }
+    public override IEnumerable<string> GetAdditionalSegment(CodeElement currentElement, string fileName)
     {
-        public RubyPathSegmenter(string rootPath, string clientNamespaceName) : base(rootPath, clientNamespaceName) { }
-        public override IEnumerable<string> GetAdditionalSegment(CodeElement currentElement, string fileName)
+        return currentElement switch
         {
-            return currentElement switch
-            {
-                CodeNamespace => new[] { fileName },// We put barrels inside namespace folders
-                _ => Enumerable.Empty<string>(),
-            };
-        }
-        public override string FileSuffix => ".rb";
-        public override string NormalizeFileName(CodeElement currentElement) => GetLastFileNameSegment(currentElement).ToSnakeCase();
-        public override string NormalizeNamespaceSegment(string segmentName) => segmentName.ToSnakeCase();
+            CodeNamespace => new[] { fileName },// We put barrels inside namespace folders
+            _ => Enumerable.Empty<string>(),
+        };
     }
+    public override string FileSuffix => ".rb";
+    public override string NormalizeFileName(CodeElement currentElement) => GetLastFileNameSegment(currentElement).ToSnakeCase();
+    public override string NormalizeNamespaceSegment(string segmentName) => segmentName.ToSnakeCase();
 }